@@ -5,7 +5,7 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build ngx-cookie --prod && ng build ngx-cookie-backend --prod",
-    "test": "ng test",
+    "test": "ng test --watch=false",
     "lint": "ng lint",
     "e2e": "ng e2e",
     "dev:ssr": "ng run backend-test-app:serve-ssr",
@@ -15,44 +15,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.0.6",
-    "@angular/common": "~9.0.6",
-    "@angular/compiler": "~9.0.6",
-    "@angular/core": "~9.0.6",
-    "@angular/forms": "~9.0.6",
-    "@angular/platform-browser": "~9.0.6",
-    "@angular/platform-browser-dynamic": "~9.0.6",
-    "@angular/platform-server": "~9.0.6",
-    "@angular/router": "~9.0.6",
-    "@nguniversal/express-engine": "^9.0.2",
+    "@angular/animations": "~10.0.8",
+    "@angular/common": "~10.0.8",
+    "@angular/compiler": "~10.0.8",
+    "@angular/core": "~10.0.8",
+    "@angular/forms": "~10.0.8",
+    "@angular/platform-browser": "~10.0.8",
+    "@angular/platform-browser-dynamic": "~10.0.8",
+    "@angular/platform-server": "~10.0.8",
+    "@angular/router": "~10.0.8",
+    "@nguniversal/express-engine": "^10.0.1",
     "express": "^4.15.2",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "rxjs": "~6.6.2",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.6",
-    "@angular-devkit/build-ng-packagr": "~0.900.6",
-    "@angular/cli": "~9.0.6",
-    "@angular/compiler-cli": "~9.0.6",
-    "@angular/language-service": "~9.0.6",
-    "@nguniversal/builders": "^9.0.2",
-    "@types/express": "^4.17.0",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "~0.1000.5",
+    "@angular-devkit/build-ng-packagr": "~0.1000.5",
+    "@angular/cli": "~10.0.5",
+    "@angular/compiler-cli": "~10.0.8",
+    "@angular/language-service": "~10.0.8",
+    "@nguniversal/builders": "^10.0.1",
+    "@types/express": "^4.17.7",
+    "@types/jasmine": "~3.5.12",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
+    "@types/node": "^14.0.27",
+    "codelyzer": "^6.0.0",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ng-packagr": "^9.0.0",
-    "protractor": "~5.4.3",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ng-packagr": "^10.0.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.10.2",
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.7"
   }
 }
