@@ -16,7 +16,7 @@ export class CookieModule {
   /**
    * Use this method in your root module to provide the CookieService
    */
-  static forRoot(options: CookieOptions = {}): ModuleWithProviders {
+  static forRoot(options: CookieOptions = {}): ModuleWithProviders<CookieModule> {
     return {
       ngModule: CookieModule,
       providers: [
@@ -30,7 +30,7 @@ export class CookieModule {
   /**
    * Use this method in your other (non root) modules to import the directive/pipe
    */
-  static forChild(options: CookieOptions = {}): ModuleWithProviders {
+  static forChild(options: CookieOptions = {}): ModuleWithProviders<CookieModule> {
     return CookieModule.forRoot(options);
   }
 }
